move detection from reader to item obj
