+++ paper2html/paper_miner.py
@@ -13,7 +13,7 @@ from paper2html.html_paper import HtmlPaper
 
 
 def read_by_extended_pdfminer(pdf_filename, line_margin_rate=None, verbose=False):
-    paper = PaperReader(line_margin_rate).read(pdf_filename)
+    paper = PaperReader().read(pdf_filename, line_margin_rate)
     if verbose:
         paper.show_layouts()
 
@@ -27,13 +27,8 @@ class PaperReader:
     """
     pdfminerを使用してpdfファイルからPaperオブジェクトを作成するクラス．
     """
-    def __init__(self, line_margin_rate=None):
+    def __init__(self):
         self.laparams = LAParams()
-        # laparams.line_margin = 0.3
-        self.laparams.boxes_flow = 1.0  # 1.0: vertical order, -1.0: horizontal order
-        # self.laparams.detect_vertical = True
-        # laparams.all_texts = True
-        self._line_margin_rate = line_margin_rate
 
     @staticmethod
     def _pdf_pages(pdf_filename):
@@ -53,12 +48,15 @@ class PaperReader:
             interpreter.process_page(page)
             yield device.get_result()
 
-    def read(self, pdf_filename):
-        if not self._line_margin_rate:
-            # TODO: line_margin設定を揃える
-            self._zap(pdf_filename)
-        else:
-            self.laparams.line_margin = self._line_margin_rate
+    def read(self, pdf_filename, line_margin_rate=None):
+        # laparams.line_margin = 0.3
+        self.laparams.boxes_flow = 1.0  # 1.0: vertical order, -1.0: horizontal order
+        # self.laparams.detect_vertical = True
+        # laparams.all_texts = True
+        self._zap(pdf_filename)
+        if line_margin_rate:
+            self.laparams.line_margin = line_margin_rate
+
         paper = Paper(self.line_height, self.line_margin)
 
         for page_number, ltpage in enumerate(self._lt_pages(pdf_filename)):
@@ -149,44 +147,6 @@ class PaperReader:
                     return horizontal_char_num < vertical_char_num
         return horizontal_char_num < vertical_char_num
 
-    def _check_separated(self, text_box):
-        mean_width = 0
-        mean_height = 0
-        tb_bbox = BBox(text_box.bbox, 'LB')
-        max_x = tb_bbox.left
-        char_count = 0
-        for child in text_box:
-            if isinstance(child, LTChar):
-                char_count += 1
-                child_bbox = BBox(child.bbox, orig='LB')
-                mean_height += child_bbox.height
-                mean_width += child_bbox.width
-                max_x = max(child_bbox.right, max_x)
-        if char_count == 0:
-            return False
-        mean_height /= char_count
-        mean_width /= char_count
-
-        last_child = None
-        for child in text_box:
-            child_bbox = BBox(child.bbox, orig='LB')
-            if isinstance(child, LTChar) and child_bbox.bottom < tb_bbox.bottom + mean_height:
-                lc_bbox = BBox(last_child.bbox, orig='LB')
-                if not last_child:
-                    last_child = child
-                elif lc_bbox.right < child_bbox.right:
-                    last_child = child
-        if last_child:
-            separated = True
-            lc_bbox = BBox(last_child.bbox, orig='LB')
-            if lc_bbox.right < max_x - mean_width:
-                separated = False
-            if last_child.get_text() == '.':
-                separated = False
-            return separated
-        else:
-            return False
-
     def _render_item(self, page, item, page_number):
         separated = False
         if isinstance(item, LTTextBoxHorizontal):
@@ -198,49 +158,17 @@ class PaperReader:
         elif isinstance(item, LTTextBoxVertical):
             item_type = PaperItemType.VTextBox
             text = item.get_text()
-            page.add_item(PaperItem(page_number, BBox(item.bbox, orig='LB'), text, item_type, separated))
+            page.add_item(PaperItem([item], page_number, BBox(item.bbox, orig='LB'), text, item_type, separated))
         else:
             print(type(item))
 
     def _render_textbox(self, page, textbox, page_number):
-        separated = False
         if self._textbox_is_vertical(textbox):
             item_type = PaperItemType.VTextBox
         else:
             item_type = PaperItemType.TextBox
-            separated = self._check_separated(textbox)
-
-        for paragraph in self._split_by_indent(textbox, item_type, separated, page_number):
-            page.add_item(paragraph)
-
-    def _split_by_indent(self, textbox, item_type, separated, page_number):
-        bbox = BBox(textbox.bbox, orig='LB')
-        split_lines = [[]]
-        lines = list(textbox)
-
-        def is_indented(line):
-            line_bbox = BBox(line.bbox, orig='LB')
-            return abs(line_bbox.bottom - bbox.bottom) > self.line_height
-        for i, line in enumerate(lines):
-            if is_indented(line):
-                if not i + 1 < len(lines):
-                    split_lines.append([])
-                elif not is_indented(lines[i + 1]):
-                    split_lines.append([])
-            split_lines[-1].append(line)
-        if len(split_lines) == 0:
-            raise ValueError('empty textbox.')
-        if len(split_lines[0]) == 0:
-            split_lines.pop(0)
-        results = []
-        for lines in split_lines:
-            bbox = BBox.unify_bboxes([BBox(line.bbox, orig='LB') for line in lines])
-            text = ''.join([line.get_text() for line in lines])
-            results.append(PaperItem(page_number, bbox, text, item_type, True))
-        if len(results) == 0:
-            raise ValueError('empty textbox.')
-        results[0].separated = separated
-        return results
+        page.add_item(PaperItem([textbox], page_number, BBox(textbox.bbox, orig='LB'),
+                                textbox.get_text(), item_type))
 
     def _render_shape(self, page, item, page_number):
         bbox = BBox(item.bbox, orig='LB')
@@ -248,7 +176,7 @@ class PaperReader:
         text = "\n\n\n\n"
         # 幅0でclipするとエラーなので膨らませておく
         bbox = bbox.inflate(1)
-        page.add_item(PaperItem(page_number, bbox, text, item_type, False))
+        page.add_item(PaperItem([item], page_number, bbox, text, item_type, False))
 
     def _render_figure(self, page, item, page_number):
         item_type = PaperItemType.Figure
@@ -263,4 +191,4 @@ class PaperReader:
                 if isinstance(child, LTChar):
                     continue
                 self._render_item(page, child, page_number)
-        page.add_item(PaperItem(page_number, BBox(item.bbox, orig='LB'), text, item_type, False))
+        page.add_item(PaperItem([item], page_number, BBox(item.bbox, orig='LB'), text, item_type, False))
